services/vertical1_tech/requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
google-generativeai==0.8.3
pydantic==2.7.1
jinja2==3.1.4
supabase==2.6.0
python-decouple==3.8
structlog==24.2.0
